+++ lib/govuk_test.rb
@@ -13,16 +13,19 @@ unless ENV["GOVUK_TEST_USE_SYSTEM_CHROMEDRIVER"]
 end
 
 module GovukTest
-  def self.configure(chrome_options: nil)
-    chrome_options ||= Selenium::WebDriver::Chrome::Options.new
-    chrome_options.headless!
-    chrome_options.add_argument("--no-sandbox") if ENV["GOVUK_TEST_CHROME_NO_SANDBOX"]
-
+  def self.configure(chrome_options: chrome_selenium_options)
     Capybara.register_driver :headless_chrome do |app|
       Capybara::Selenium::Driver.new(app, browser: :chrome, options: chrome_options)
     end
 
     Capybara.javascript_driver = :headless_chrome
     Capybara.server = :puma, { Silent: true }
+  end
+
+  def self.chrome_selenium_options
+    chrome_options = Selenium::WebDriver::Chrome::Options.new
+    chrome_options.headless!
+    chrome_options.add_argument("--no-sandbox") if ENV["GOVUK_TEST_CHROME_NO_SANDBOX"]
+    chrome_options
   end
 end